feat: AddDetailPlanRequest 클래스 PK 재설정

+++ src/main/java/com/ssafy/tourmates/client/controller/dto/tripPlan/AddDetailPlanRequest.java
@@ -11,5 +11,5 @@ public class AddDetailPlanRequest {
 
     @NotNull
     @Size(min = 1)
-    List<Long> contendIds;
+    List<Integer> contendIds;
 }
